Fixed me being dumb
Link issues due to not creating my headers properly
Can once again build a release build, yay!

+++ deploy/src/movement_controller.h
@@ -6,7 +6,7 @@
 #define MOVEMENT_CONTROLLER_H
 
 #include <memory>
-
+#include <cstdint>
 
 #include "motor_driver.h"
 
@@ -28,25 +28,25 @@ public:
 
 private:
   std::unique_ptr<PCA9685> pwm_driver;
-  const int pwm_left_pin = 13;
-  const int in1_left_pin = 11;
-  const int in2_left_pin = 12;
+  const uint8_t pwm_left_pin = 13;
+  const uint8_t in1_left_pin = 11;
+  const uint8_t in2_left_pin = 12;
 
-  const int pwm_right_pin = 2;
-  const int in1_right_pin = 4;
-  const int in2_right_pin = 3;
+  const uint8_t pwm_right_pin = 2;
+  const uint8_t in1_right_pin = 4;
+  const uint8_t in2_right_pin = 3;
 
   // Values for calculating a smooth transition
   // value can vary do from 0 to 51
-  uint32_t current_value = 25;
-  uint32_t target_value = 25;
-  const uint32_t step = 1;
+  uint16_t current_value = 25;
+  uint16_t target_value = 25;
+  const uint16_t step = 1;
 
   // Manual Control toggle
   int auto_control = true;
   int last_direction = 0;
-  int current_left_pwm = 0;
-  int current_right_pwm = 0;
+  uint16_t current_left_pwm = 0;
+  uint16_t current_right_pwm = 0;
 };
